ui: moved GitAliasMap from components folder to views folder because it is not a reusable component
